% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{control_args}
\alias{control_args}
\title{Prepare control arguments}
\usage{
control_args(args)
}
\arguments{
\item{args}{A list of arguments. See See *Details*.}
}
\value{
A list of arguments
}
\description{
Prepare control arguments
}
\details{
The following arguments are supported:
\describe{
\item{\code{max_rounds}}{The maximum number of rounds of coordinate descent (default: 50.)}
\item{\code{max_iter}}{The maximum number of iterations per round (default: 1000.)}
\item{\code{rel_tol}}{The relative tolerance for convergence (default: \code{1e-6}.)}
\item{\code{verbose}}{The verbosity level. Must be 0, 1(default), 2, or 3.}
\item{\code{update_method}}{The update method. Must be \code{"PGD"} or \code{"BPGD"}. Currently overwritten.}
\item{\code{backtracking_fraction}}{The backtracking fraction. Must be between 0 and 1 (default: \code{0.9}.)}
\item{\code{...}}{Additional arguments. Currently ignored.}
}
}
