% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate synthetic data}
\usage{
generate_synthetic_data(
  n_subjects = 100,
  n_timepoints = 11,
  prop_observed = 0.7,
  observation_variance = 1,
  random_effect_variance_ratio = 1,
  random_effect_ar1_correlation = 0.9,
  effect_size = 1,
  seed = 1
)
}
\arguments{
\item{seed}{}
}
\description{
Generate synthetic data
}
