% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_synthetic_data_p}
\alias{generate_synthetic_data_p}
\title{Generate synthetic data (multiple features)}
\usage{
generate_synthetic_data_p(
  n_subjects = 100,
  n_timepoints = 11,
  n_features = 2,
  feature_type = "binary",
  observation_variance = 1,
  random_effect_variance_ratio = 1,
  random_effect_ar1_correlation = 0.9,
  effect_size = 1,
  effect_groupsparsity = 0.5,
  effect_sparsity = 0.5,
  prop_observed = 0.7,
  missingness = "uniform",
  seed = 1
)
}
\arguments{
\item{n_subjects}{Number of subjects (default 100).}

\item{n_timepoints}{Number of time points at which the data is sampled spread regularly on the \eqn{[0,1]} interval (default 11).}

\item{n_features}{Number of features in addition to the intercept (default 2).}

\item{feature_type}{Type of feature (implemented: "binary" (default), "uniform", "normal").}

\item{observation_variance}{Variance of the observation error (default 1).}

\item{random_effect_variance_ratio}{Ratio of the variance of the random effect to the variance of the observation error (default 1).}

\item{random_effect_ar1_correlation}{Correlation between two random effects at two consecutive timepoints (default 0.9).}

\item{effect_size}{Effect size of the features (default 1).}

\item{effect_groupsparsity}{Proportion of features that have no effect (default 0.5).}

\item{effect_sparsity}{Proportion of the domain that has null effect (default 0.5).}

\item{prop_observed}{Proportion of observed time points (default 0.7).}

\item{missingness}{Missing observation mechanism (implemented: "uniform" (default), "contiguous", "sqrt", "fixed_uniform")}

\item{seed}{Seed for the random number generator (default 1).}
}
\value{
A list containing the following elements:
\describe{
\item{data}{A data frame containing the longitudinal data in long format.}
\item{data_wide}{A data frame containing the longitudinal data in wide format.}
\item{data_wide_imputed}{A data frame containing the longitudinal data in wide format with imputed missing values.}
\item{true_values}{A data frame containing the true values of the fixed effects.}
\item{times}{A vector of timepoints at which the data is sampled.}
\item{colnames}{A list containing the column names of the data frames.}
}
}
\description{
Generate synthetic data (multiple features)
}
\details{
The data is generated according to the following model:
\deqn{y_{ij} = \beta_0(t_{ij}) + \sum_{k=1}^{p} \beta_k(t_{ij}) x_{ik} + \theta_{ij} + \epsilon_{ij}}
where \eqn{y_{ij}} is the \eqn{j}th response of subject \eqn{i}, at time \eqn{(t_{ij})},
\eqn{x_{ik}} is the \eqn{k}th feature of subject \eqn{i},
\eqn{\theta_{ij}} is the random effect of subject \eqn{i} at time \eqn{j},
\eqn{\epsilon_{ij}} is the observation error of subject \eqn{i} at time \eqn{j},
\eqn{\beta_0(\cdot)} is the intercept function, and \eqn{\beta_1(\cdot)} is the effect size.
The random effects are generated according to a multivariate normal distribution
with mean 0 and covariance matrix \eqn{\Sigma} such that \eqn{\Sigma_{ij} = r\sigma^2\rho^{|i-j|}}
(\eqn{\rho} defined by \code{random_effect_ar1_correlation} and \eqn{r} by \code{random_effect_variance_ratio}).
The observation errors are generated according to a multivariate normal distribution
with mean 0 and covariance matrix \eqn{\sigma^2 I}
(\eqn{\sigma^2} defined by \code{observation_variance}).
The random effects and observation errors are independent.
The observations are then sampled uniformly at a proportion \code{prop_observed} of the timepoints.
An imputed version of the data is also provided where the missing values are imputed using
the function [refund::fpca.sc].
The data is returned in three formats:
\describe{
\item{data}{A data frame containing the longitudinal data in long format.}
\item{data_wide}{A data frame containing the longitudinal data in wide format.}
\item{data_wide_imputed}{A data frame containing the longitudinal data in wide format with imputed missing values.}
}
}
\examples{
instance = generate_synthetic_data_p()

}
