% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{kernel_args}
\alias{kernel_args}
\title{Prepare kernel arguments}
\usage{
kernel_args(args)
}
\arguments{
\item{args}{A list of arguments. See *Details*.}
}
\value{
A list of arguments
}
\description{
Prepare kernel arguments
}
\details{
The following arguments are supported:
\describe{
\item{\code{name}}{The kernel function. Currently, only \code{"gaussian"} is supported.}
\item{\code{scale}}{The scale parameter. If \code{NULL}, a logarithmic grid search is used.}
\item{\code{n_scale}}{The number of scale parameters to use if \code{scale} is \code{NULL} (default.)}
\item{\code{...}}{Additional arguments. Currently ignored.}
}

For the grid search, the range is defined to be \eqn{[L, U]} where
\eqn{L} is half the smallest gap in estimated time points and
\eqn{U} is twice the range of estimated time points.
}
