% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSVCMM-package.R
\docType{package}
\name{LSVCMM-package}
\alias{LSVCMM}
\alias{LSVCMM-package}
\title{LSVCMM: Locally Sparse Varying Coefficients Mixed Model}
\description{
Implements a locally sparse varying coefficient mixed model proposed by <redacted>.
}
\author{
\strong{Maintainer}: Redacted Redacted \email{redacted@redacted.com}

}
\keyword{internal}
