% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsvcmm.R
\name{lsvcmm}
\alias{lsvcmm}
\title{Title}
\usage{
lsvcmm(
  data,
  response,
  subject,
  time,
  vc_covariates = NULL,
  nvc_covariates = NULL,
  offset = NULL,
  add_intercept = T,
  estimated_time = NULL,
  family = DEFAULT_FAMILY_ARGS,
  kernel = DEFAULT_KERNEL_ARGS,
  penalty = DEFAULT_PENALTY_ARGS,
  working_covariance = DEFAULT_WORKING_COVARIANCE_ARGS,
  control = DEFAULT_CONTROL_ARGS
)
}
\arguments{
\item{control}{}
}
\description{
Title
}
