% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{working_covariance_args}
\alias{working_covariance_args}
\title{Prepare working covariance arguments}
\usage{
working_covariance_args(args)
}
\arguments{
\item{args}{A list of arguments. See See *Details*.}
}
\value{
A list of arguments
}
\description{
Prepare working covariance arguments
}
\details{

}
