% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{family_args}
\alias{family_args}
\title{Prepare family arguments}
\usage{
family_args(args)
}
\arguments{
\item{args}{A list of arguments. See See *Details*.}
}
\value{
A list of arguments
}
\description{
Prepare family arguments
}
\details{
The following arguments are supported:
\describe{
 \item{\code{response}}{The response distribution. Currently, only \code{"gaussian"} is supported.}
 \item{\code{link}}{The link function. Currently, only \code{"identity"} is supported.}
 \item{\code{...}}{Additional arguments. Currently ignored.}
}

If an argument is not specified, the default value is used.
}
