% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSVCMM-package.R
\docType{package}
\name{LSVCMM-package}
\alias{LSVCMM}
\alias{LSVCMM-package}
\title{LSVCMM: Locally Sparse Varying Coefficients Mixed Model}
\description{
Implements a locally sparse varying coefficient mixed model proposed by Fontaine et al. (2023+).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fontaine618.github.io/LSVCMM}
  \item Report bugs at \url{https://github.com/fontaine618/LSVCMM/issues}
}

}
\author{
\strong{Maintainer}: Simon Fontaine \email{simfont@umich.edu}

}
\keyword{internal}
