% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{lsvcmm.boot}
\alias{lsvcmm.boot}
\title{Title}
\usage{
lsvcmm.boot(
  data,
  response,
  subject,
  time,
  vc_covariates = NULL,
  nvc_covariates = NULL,
  offset = NULL,
  add_intercept = T,
  estimated_time = NULL,
  family = DEFAULT_FAMILY_ARGS,
  kernel = DEFAULT_KERNEL_ARGS,
  penalty = DEFAULT_PENALTY_ARGS,
  working_covariance = DEFAULT_WORKING_COVARIANCE_ARGS,
  control = DEFAULT_CONTROL_ARGS,
  n_samples = 1000
)
}
\arguments{
\item{n_samples}{}
}
\description{
Title
}
