% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{working_covariance_args}
\alias{working_covariance_args}
\title{Prepare working covariance arguments}
\usage{
working_covariance_args(args)
}
\arguments{
\item{args}{A list of arguments. See See *Details*.}
}
\value{
A list of arguments
}
\description{
Prepare working covariance arguments
}
\details{
The following arguments are supported:
\describe{
\item{\code{name}}{The working covariance function. Currently, only \code{"compound_symmetry"} is supported.}
\item{\code{estimate}}{Whether to estimate the parameters. Must be \code{TRUE} (default) or \code{FALSE}.}
\item{\code{ratio}}{The ratio of the variance of the random effects to the variance of the error term.
If negative (default), the ratio is estimated to `log(n)`.}
\item{\code{correlation}}{The correlation for AR(1).}
}
}
